Add anchored behavior to RCTScrollView.
By setting `anchorMode=true`, the scroll view tracks an anchor element
as the user scrolls and ensures that the scroll view maintains its position
(such that the anchor element is in the same place on the screen) as items
are added and removed from the view.

Very useful for implementing a chat stream where both old and new messages
are fetched on demand as the user scrolls.